requirements.txt
Flask==1.0.2
facebook-sdk==2.0.0
gunicorn==19.8.1
